Extract native request catalog and gateway boundary ownership
Move request decoding, model catalog construction, and intact gateway critical sections into dedicated modules without changing the C ABI or engine lifetime scopes.
