+++ packages/agent-native-bridge/agent-native-bridge.cabal
@@ -75,6 +75,9 @@ foreign-library haskell-agent-bridge
                     , Agent.CLI.MacOS.RepositoryChecks
                     , Agent.CLI.MacOS.RepositoryDeliveryBridge
                     , Agent.CLI.MacOS.RepositoryReviewBridge
+                    , Agent.CLI.MacOS.NativeRequest
+                    , Agent.CLI.MacOS.NativeGatewayBoundary
+                    , Agent.CLI.MacOS.NativeModelCatalog
                     , Agent.CLI.MacOS.RepositoryInput
                     , Agent.CLI.MacOS.ResourceAdmin
     c-sources:        cbits/HaskellAgentBridgeRuntime.c
@@ -159,6 +162,9 @@ test-suite agent-native-bridge-test
                      , Agent.CLI.MacOS.RepositoryChecks
                      , Agent.CLI.MacOS.RepositoryDeliveryBridge
                      , Agent.CLI.MacOS.RepositoryReviewBridge
+                     , Agent.CLI.MacOS.NativeRequest
+                     , Agent.CLI.MacOS.NativeGatewayBoundary
+                     , Agent.CLI.MacOS.NativeModelCatalog
                     , Agent.CLI.MacOS.ResourceAdmin
                      , Agent.CLI.MacOS.BrowserBridgeFFISpec
                      , Agent.CLI.MacOS.ComputerBridgeSpec
