+++ packages/agent-native-bridge/ffi/Agent/CLI/MacOS/NativeRequest.hs
@@ -1,0 +1,156 @@
+-- | The JSON request vocabulary accepted by the native engine.
+-- Decoding lives independently of FFI handles and worker lifetimes.
+module Agent.CLI.MacOS.NativeRequest
+    ( BridgeRequest(..)
+    , TurnStart(..)
+    , TurnReference(..)
+    , ApprovalResolution(..)
+    , SessionPageRequest(..)
+    , SessionReference(..)
+    , ModelsListRequest(..)
+    , parseParams
+    , turnStartCleanupId
+    ) where
+
+import Agent.CLI.Options (parseEffort)
+import Data.Aeson ((.:), (.:?))
+import qualified Data.Aeson as Aeson
+import qualified Data.Aeson.Types as Aeson
+import Data.Int (Int64)
+import Data.Maybe (fromMaybe)
+import Data.Text (Text)
+import qualified Data.Text as Text
+
+data BridgeRequest = BridgeRequest
+    { requestId :: !Text
+    , requestMethod :: !Text
+    , requestParams :: !Aeson.Value
+    }
+
+instance Aeson.FromJSON BridgeRequest where
+    parseJSON = Aeson.withObject "BridgeRequest" \object ->
+        BridgeRequest
+            <$> object .: "id"
+            <*> object .: "method"
+            <*> (object .:? "params" Aeson..!= Aeson.object [])
+
+data TurnStart = TurnStart
+    { turnStartId :: !Text
+    , turnStartPrompt :: !Text
+    , turnStartSessionId :: !(Maybe Text)
+    , turnStartCwd :: !FilePath
+    , turnStartProvider :: !(Maybe Text)
+    , turnStartModel :: !(Maybe Text)
+    , turnStartEffort :: !(Maybe Text)
+    , turnStartWorktree :: !Bool
+    , turnStartComputerUse :: !Bool
+    }
+
+instance Aeson.FromJSON TurnStart where
+    parseJSON = Aeson.withObject "TurnStart" \object -> do
+        turnStartId <- object .: "turnId"
+        turnStartPrompt <- object .: "prompt"
+        turnStartSessionId <- object .:? "sessionId"
+        turnStartCwd <- object .: "cwd"
+        turnStartProvider <- object .:? "provider"
+        turnStartModel <- object .:? "model"
+        turnStartEffort <- object .:? "effort"
+        _ <- traverse
+            (either fail pure . parseEffort)
+            turnStartEffort
+        turnStartWorktree <- object .:? "worktree" Aeson..!= False
+        turnStartComputerUse <-
+            object .:? "computerUse" Aeson..!= False
+        let start = TurnStart
+                { turnStartId
+                , turnStartPrompt
+                , turnStartSessionId
+                , turnStartCwd
+                , turnStartProvider
+                , turnStartModel
+                , turnStartEffort
+                , turnStartWorktree
+                , turnStartComputerUse
+                }
+        case
+            ( turnStartSessionId
+            , turnStartWorktree
+            , turnStartProvider
+            , turnStartModel
+            )
+          of
+            (Just _, True, _, _) ->
+                fail "a worktree can only be created for a new session"
+            (_, _, Nothing, Nothing) -> pure start
+            (_, _, Just _, Just _) -> pure start
+            _ -> fail "provider and model must be supplied together"
+
+turnStartCleanupId :: Text -> Aeson.Value -> Text
+turnStartCleanupId requestId params =
+    fromMaybe requestId $
+        Aeson.parseMaybe
+            (Aeson.withObject "TurnStartCleanup" (.:? "turnId"))
+            params
+            >>= id
+            >>= nonBlank
+  where
+    nonBlank value
+        | Text.null (Text.strip value) = Nothing
+        | otherwise = Just value
+
+data TurnReference = TurnReference
+    { turnReferenceId :: !Text
+    }
+
+instance Aeson.FromJSON TurnReference where
+    parseJSON = Aeson.withObject "TurnReference" \object ->
+        TurnReference <$> object .: "turnId"
+
+data ApprovalResolution = ApprovalResolution
+    { approvalResolutionId :: !Text
+    , approvalResolutionDecision :: !Text
+    }
+
+instance Aeson.FromJSON ApprovalResolution where
+    parseJSON = Aeson.withObject "ApprovalResolution" \object ->
+        ApprovalResolution
+            <$> object .: "approvalId"
+            <*> object .: "decision"
+
+data SessionPageRequest = SessionPageRequest
+    { sessionPageId :: !Text
+    , sessionPageBefore :: !(Maybe Int64)
+    , sessionPageLimit :: !(Maybe Int)
+    }
+
+instance Aeson.FromJSON SessionPageRequest where
+    parseJSON = Aeson.withObject "SessionPageRequest" \object ->
+        SessionPageRequest
+            <$> object .: "id"
+            <*> object .:? "before"
+            <*> object .:? "limit"
+
+data SessionReference = SessionReference
+    { sessionReferenceId :: !Text
+    }
+
+instance Aeson.FromJSON SessionReference where
+    parseJSON = Aeson.withObject "SessionReference" \object ->
+        SessionReference <$> object .: "id"
+
+data ModelsListRequest = ModelsListRequest
+    { modelsListCwd :: !FilePath
+    , modelsListSessionId :: !(Maybe Text)
+    }
+
+instance Aeson.FromJSON ModelsListRequest where
+    parseJSON = Aeson.withObject "ModelsListRequest" \object ->
+        ModelsListRequest
+            <$> object .: "cwd"
+            <*> object .:? "sessionId"
+
+parseParams :: Aeson.FromJSON value => BridgeRequest -> Either Text value
+parseParams request =
+    case Aeson.parseEither Aeson.parseJSON request.requestParams of
+        Left err -> Left (Text.pack err)
+        Right value -> Right value
